refactor: separate auth service from user service
